requirements.txt
streamlit
pandas
sqlalchemy
pymysql
psycopg2-binary
st-pages